pyproject.toml
[project]
name = "django-chat"
version = "0.1.0"
description = "Django Chat Application"
authors = [
    { name = "hckim", email = "hckim@example.com" }
]
dependencies = [
    "django>=5.0.0",
    "channels>=4.0.0",
    "psycopg>=3.1.18",
    "redis>=5.0.1",
    "channels-redis>=4.2.0",
    "django-environ>=0.11.2",
    "djangorestframework>=3.15.2",
    "daphne>=4.1.0",
    "drf-nested-routers>=0.93.5",
    "django-redis>=5.4.0",
    "djangorestframework-simplejwt>=5.5.0",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chat"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
]
